+++ packages/e2e-tests/specs/experiments/multi-entity-editing.test.js
@@ -124,6 +124,18 @@ const openEntitySavePanel = async () => {
 		await page.waitForSelector( '.entities-saved-states__panel' );
 	}
 	// If we made it this far, the panel is opened.
+
+	// Expand to view savable entities if necessary.
+	const reviewChangesButton = await page.$(
+		'.entities-saved-states__review-changes-button'
+	);
+	const [ needsToOpen ] = await reviewChangesButton.$x(
+		'//*[contains(text(),"Review changes.")]'
+	);
+	if ( needsToOpen ) {
+		await reviewChangesButton.click();
+	}
+
 	return true;
 };
