+++ packages/editor/src/components/entities-saved-states/entity-record-item.js
@@ -1,0 +1,77 @@
+/**
+ * WordPress dependencies
+ */
+import { CheckboxControl, Button, PanelRow } from '@wordpress/components';
+import { __ } from '@wordpress/i18n';
+import { useSelect, useDispatch } from '@wordpress/data';
+import { useCallback } from '@wordpress/element';
+
+export default function EntityRecordItem( {
+	record,
+	checked,
+	onChange,
+	closePanel,
+} ) {
+	const { name, kind, title, key } = record;
+	const parentBlockId = useSelect( ( select ) => {
+		// Get entity's blocks.
+		const { blocks = [] } = select( 'core' ).getEditedEntityRecord(
+			kind,
+			name,
+			key
+		);
+		// Get parents of the entity's first block.
+		const parents = select( 'core/block-editor' ).getBlockParents(
+			blocks[ 0 ]?.clientId
+		);
+		// Return closest parent block's clientId.
+		return parents[ parents.length - 1 ];
+	}, [] );
+
+	const isSelected = useSelect(
+		( select ) => {
+			const selectedBlockId = select(
+				'core/block-editor'
+			).getSelectedBlockClientId();
+			return selectedBlockId === parentBlockId;
+		},
+		[ parentBlockId ]
+	);
+	const isSelectedText = isSelected ? __( 'Selected' ) : __( 'Select' );
+	const { selectBlock } = useDispatch( 'core/block-editor' );
+	const selectParentBlock = useCallback( () => selectBlock( parentBlockId ), [
+		parentBlockId,
+	] );
+	const selectAndDismiss = useCallback( () => {
+		selectBlock( parentBlockId );
+		closePanel();
+	}, [ parentBlockId ] );
+
+	return (
+		<PanelRow>
+			<CheckboxControl
+				label={ <strong>{ title || __( 'Untitled' ) }</strong> }
+				checked={ checked }
+				onChange={ onChange }
+			/>
+			{ parentBlockId ? (
+				<>
+					<Button
+						onClick={ selectParentBlock }
+						className="entities-saved-states__find-entity"
+						disabled={ isSelected }
+					>
+						{ isSelectedText }
+					</Button>
+					<Button
+						onClick={ selectAndDismiss }
+						className="entities-saved-states__find-entity-small"
+						disabled={ isSelected }
+					>
+						{ isSelectedText }
+					</Button>
+				</>
+			) : null }
+		</PanelRow>
+	);
+}
